Bug fix TodoList

+++ src/components/ToDo.js
@@ -2,10 +2,9 @@ import { dbService } from "fBase";
 import { deleteDoc, doc, updateDoc } from "firebase/firestore";
 import React, { useState } from "react";
 
-const ToDo = ({ toDoObj, isOwner }) => {
+const ToDo = ({ toDoObj }) => {
   const [editToggle, setEditToggle] = useState(false);
   const [isEdit, setIsEdit] = useState(false);
-  const [selectedToDo, setSelectedToDo] = useState("");
   const [newTask, setNewTask] = useState(toDoObj.task);
   const [isChecked, setIsChecked] = useState(true); // check box value 초기 값이 true여야지 첫 번쨰 클릭때 onCheck 함수에서 true로 바뀜
   const toDoRef = doc(dbService, "todos", `${toDoObj.id}`);
@@ -24,8 +23,12 @@ const ToDo = ({ toDoObj, isOwner }) => {
   };
 
   const onDeleteToDo = async () => {
-    setEditToggle(false);
-    await deleteDoc(toDoRef);
+    const ok = window.confirm("정말로 삭제할까요?");
+    if (ok) {
+      setEditToggle(false);
+      await deleteDoc(toDoRef);
+    }
+
   };
 
   const onEditToDo = async (event) => {
@@ -49,17 +52,31 @@ const ToDo = ({ toDoObj, isOwner }) => {
     });
   };
 
+  const onCancle = (prev) => setIsEdit((prev) => !prev);
+
 
   return (
     <>
-      {isEdit && <span>asd</span>}
-      <div>
+      <div >
         <input type="checkbox"
           value={isChecked}
           onChange={onCheck}
           checked={toDoObj.done}
         />
-        <li>{toDoObj.task}</li>
+        {isEdit ?
+          (<li>
+            <form onSubmit={onEditSubmit}>
+              <input
+                value={newTask}
+                onChange={onEditChange}
+              />
+              <button name="edit" type="submit">수정!</button>
+              <button name="cancle" type="button" onClick={onCancle} >취소</button>
+            </form>
+          </li>) :
+          <li>{toDoObj.task}</li>
+        }
+
         <button
           value="edit"
           onClick={onEditClick}>...</button>
